\name{mfrmmx-package}
\alias{mfrmmx-package}
\alias{mfrmmx}
\docType{package}
\title{\packageTitle{mfrmmx}}
\description{\packageDescription{mfrmmx}}
\section{Package Content}{\packageIndices{mfrmmx}}
\author{\packageAuthor{mfrmmx}}
\section{Maintainer}{\packageMaintainer{mfrmmx}}
\keyword{package}
